# swagger.json
{
    "openapi": "3.0.0",
    "paths": {
        "/agreement": {
            "post": {
                "operationId": "AgreementController_saveAgreement",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAgreementDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "agreement"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "get": {
                "operationId": "AgreementController_getAllAgreement",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "agreement"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/agreement/{id}": {
            "get": {
                "operationId": "AgreementController_getAgreementById",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "agreement"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/agreement/update/{id}": {
            "put": {
                "operationId": "AgreementController_updateAgreementById",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAgreementDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "agreement"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/offer/provider/{id}": {
            "post": {
                "operationId": "OfferController_create",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOfferDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "offer"
                ]
            },
            "get": {
                "operationId": "OfferController_getOfferByProviderId",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "offer"
                ]
            }
        },
        "/offer/agreement/{id}": {
            "get": {
                "operationId": "OfferController_getAllOfferByAgreementId",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "offer"
                ]
            }
        },
        "/offer/{id}": {
            "get": {
                "operationId": "OfferController_getOfferById",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "offer"
                ]
            }
        },
        "/offer/update/{id}": {
            "put": {
                "operationId": "OfferController_updateOfferById",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOfferDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "offer"
                ]
            }
        },
        "/materialGroup": {
            "post": {
                "operationId": "MaterialGroupController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMaterialGroupDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "materialGroup"
                ]
            },
            "get": {
                "operationId": "MaterialGroupController_getAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "materialGroup"
                ]
            }
        },
        "/materialGroup/update/{id}": {
            "put": {
                "operationId": "MaterialGroupController_updateOfferById",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMaterialGroupDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "materialGroup"
                ]
            }
        },
        "/auth/signup": {
            "post": {
                "operationId": "AuthController_signup",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignupDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "authentication"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/auth/signin": {
            "post": {
                "operationId": "AuthController_signin",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SigninDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "authentication"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/provider/auth/signup": {
            "post": {
                "operationId": "AuthProviderController_signup",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProviderSignupDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "authentication-provider"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/provider/auth/signin": {
            "post": {
                "operationId": "AuthProviderController_signin",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SigninDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "authentication-provider"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "info": {
        "title": "Provider-management-platform",
        "description": "API to handle everything related to provider-management",
        "version": "1.0",
        "contact": {}
    },
    "tags": [],
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
            "DeadlineDto": {
                "type": "object",
                "properties": {
                    "deadline": {
                        "type": "string"
                    },
                    "teamDeadline": {
                        "type": "string"
                    },
                    "contractDeadline": {
                        "type": "string"
                    }
                },
                "required": [
                    "deadline",
                    "teamDeadline",
                    "contractDeadline"
                ]
            },
            "CreateAgreementDto": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "number"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "skills": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "validFrom": {
                        "type": "string"
                    },
                    "validUntil": {
                        "type": "string"
                    },
                    "cycle": {
                        "type": "string"
                    },
                    "materialGroup": {
                        "type": "string"
                    },
                    "dailyRateIndicator": {
                        "type": "string"
                    },
                    "deadline": {
                        "$ref": "#/components/schemas/DeadlineDto"
                    }
                },
                "required": [
                    "userId",
                    "title",
                    "description",
                    "skills",
                    "validFrom",
                    "validUntil",
                    "cycle",
                    "materialGroup",
                    "dailyRateIndicator",
                    "deadline"
                ]
            },
            "CreateOfferDto": {
                "type": "object",
                "properties": {
                    "agreementId": {
                        "type": "number"
                    },
                    "skills": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "experienceLevel": {
                        "type": "string"
                    },
                    "salary": {
                        "type": "string"
                    }
                },
                "required": [
                    "agreementId",
                    "skills",
                    "experienceLevel",
                    "salary"
                ]
            },
            "CreateMaterialGroupDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string"
                    }
                },
                "required": [
                    "name",
                    "description",
                    "price"
                ]
            },
            "SignupDto": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "firstName",
                    "lastName",
                    "email",
                    "password"
                ]
            },
            "SigninDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
            "ProviderSignupDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "existsSince": {
                        "type": "string"
                    },
                    "validFrom": {
                        "type": "string"
                    },
                    "validUntil": {
                        "type": "string"
                    },
                    "masterAgreementType": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "password",
                    "name",
                    "address",
                    "existsSince",
                    "validFrom",
                    "validUntil",
                    "masterAgreementType"
                ]
            }
        }
    }
}